Split PaymentModule EventPresenter
