+++ app/AccountancyModule/PaymentModule/templates/AddEventParticipants/default.latte
@@ -1,3 +1,6 @@
+{varType Model\DTO\Payment\Group $group}
+{varType bool $showForm}
+
 {block #content}
 {include '../@groupBackLink.latte', group => $group}
