+++ app/AccountancyModule/PaymentModule/templates/GroupList/default.latte
@@ -38,7 +38,7 @@
 
                         <div class="dropdown-divider"></div>
 
-                        <a n:href="Event:default" class="dropdown-item">
+                        <a n:href="SelectEventForGroup:" class="dropdown-item">
                             <span class="mr-1"><i class="fas fa-calendar"></i></span> Akce
                             <span class="small d-block">Umožnuje načíst účastnické poplatky.</span>
                         </a>
