+++ app/AccountancyModule/PaymentModule/templates/SelectEventForGroup/default.latte
@@ -1,3 +1,5 @@
+{varType Model\Event\Event[] $events}
+
 {block content}
 <h1 n:inner-block="title">Nová skupina plateb pro akci</h1>
